Add support for loading Turkish translations on server side requests
Need to read more in order to figure out how to use it properly for the
client side since it seems to be hard coding the `guides` path.

Also worth nothing that a new Guide instance is created on each request
for now but we should probably fix it so that an instance is reused.